backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
psycopg2-binary
python-dotenv
pydantic>=2
python-multipart
passlib[bcrypt]
python-jose[cryptography]
